tests/system/requirements.txt
# Python dependencies for Lumos system tests

pytest>=7.0.0
pytest-timeout>=2.1.0
pyyaml>=6.0
